moved FreezerView.vue form under view directory of form.
