+++ client/src/forms/view/FreezerView.vue
